chore(deps): bump the patch group across 1 directory with 17 updates
Bump patch dependencies with fixes for MSRV compatibility:
- Update anyhow, chrono, libc, log, serde_json, thiserror, tracing,
  cfg-if, futures, tokio-stream, serde_bytes, caps, dbus, tracing-subscriber
- Skip time (0.3.47 requires rustc 1.88.0) and tokio-util (0.7.18
  has private API issue with tokio) to maintain rustc 1.86.0 compat
- Fix clippy warnings: needless_borrows_for_generic_args, unnecessary_cast,
  bool_assert_comparison

Originally from dependabot PR #1087.

+++ crates/containerd-shim-wasm/src/testing.rs
@@ -260,7 +260,7 @@ impl<WasiEngine: Shim> WasiTest<WasiEngine> {
 
     pub fn kill(&self) -> Result<&Self> {
         log::info!("sending SIGKILL");
-        self.instance.kill(SIGKILL as u32).block_on()?;
+        self.instance.kill(SIGKILL).block_on()?;
         Ok(self)
     }
